CuBa-Desktop-SW Matlab

+++ 2_1D_Modell/2_5_BBB_SW/CuBa_Desktop/lib/Global.h
@@ -5,11 +5,11 @@
 
 using Int8			= signed char;
 using Int16			= signed short;
-using Int32			= signed long;
+using Int32			= signed int;
 using Int64			= signed long long;
 using UInt8			= unsigned char;
 using UInt16		= unsigned short;
-using UInt32 		= unsigned long;
+using UInt32 		= unsigned int;
 using UInt64		= unsigned long long;
 using Float32		= float;
 using Double64		= double;
